Website/requirements.txt
scikit-learn==1.2.0
scipy==1.10.0
streamlit==1.22.0
streamlit-chat==0.0.2.2
requests~=2.30.0
google-generativeai
